Safari-611.1.10/LayoutTests/canvas/philip/tests/fallback.nested.html

<!DOCTYPE html>
<title>Canvas test: fallback.nested</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><canvas><p class="fallback">FAIL (fallback content)</p></canvas><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

_assertEqual(canvas.childNodes.length, 2, "canvas.childNodes.length", "2");


});
</script>
